implement hint page using mock data
